Add hack to stop regeneration of .inc file needlessly.
THIS IS NOT A PROPER SOLUTION:

A correct fix is, for example, to move the source
generation to its own toplevel place so it can be
processed before any code that uses it.

The current setup is problematic because third party
code includes the generated header causing there
to be a cyclic dependency between the two.